% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_output,ANY-method}
\alias{format_output,ANY-method}
\title{Default to format output}
\usage{
\S4method{format_output}{ANY}(output, compartments, reduce_age = FALSE)
}
\arguments{
\item{output}{output of simulate}

\item{compartments}{list of compartments you want to return}

\item{reduce_age}{logical, if TRUE, summarise the age groups}
}
\description{
Default to format output
}
