% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_output}
\alias{format_output}
\title{Format the output of simulate}
\usage{
format_output(output, compartments, reduce_age = FALSE)
}
\arguments{
\item{output}{output of simulate}

\item{compartments}{list of compartments you want to return}

\item{reduce_age}{logical, if TRUE, summarise the age groups}
}
\value{
data frame of results
}
\description{
Format the output of simulate
}
