% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_point_estimate}
\alias{project_point_estimate}
\title{Function to project immunity using point estimates of the parameters}
\usage{
project_point_estimate(
  type,
  t_projection_starts,
  t_projection_ends,
  age_group_sizes,
  death_rates,
  tt_death_rates = NULL,
  birth_rates,
  tt_birth_rates = NULL,
  duration_of_maternal_immunity,
  vaccine_names,
  force_of_infection,
  tt_force_of_infection = NULL,
  vaccine_efficacy,
  vaccine_doses,
  tt_vaccine_doses = NULL,
  duration_of_immunity,
  S_0,
  R_0,
  M_0
)
}
\description{
Runs multiple models for each vaccine type (determined by the first dimension of relevant parameters
}
