% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_parameters.R
\name{sample_parameters}
\alias{sample_parameters}
\title{Function to sample R0, duration of infectiousness and pre-infectioussness for a given disease}
\usage{
sample_parameters(disease, n_samples = NULL)
}
\arguments{
\item{disease}{Which disease to sample}

\item{n_samples}{number of samples to draw, if NULL takes the central value}
}
\description{
Function to sample R0, duration of infectiousness and pre-infectioussness for a given disease
}
