% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate}
\alias{simulate}
\title{Run model}
\usage{
simulate(
  type,
  t,
  age_group_sizes,
  death_rates,
  tt_death_rates = NULL,
  birth_rates,
  tt_birth_rates = NULL,
  force_of_infection,
  tt_force_of_infection = NULL,
  vaccine_efficacy,
  vaccine_doses,
  tt_vaccine_doses = NULL,
  duration_of_immunity,
  duration_of_maternal_immunity,
  S_0,
  R_0,
  M_0,
  additional_parameters = NULL
)
}
\arguments{
\item{type}{Type of odin model to run}

\item{t}{Time-series of times to collect data from}

\item{age_group_sizes}{Sizes of each age group}

\item{death_rates}{Death rates for each age group}

\item{tt_death_rates}{Times at which death rates change}

\item{birth_rates}{Birth rates for each age group}

\item{tt_birth_rates}{Times at which birth rates change}

\item{force_of_infection}{Force of infection for each age group}

\item{tt_force_of_infection}{Times at which force of infection changes}

\item{vaccine_efficacy}{Efficacy of vaccine}

\item{vaccine_doses}{Doses of vaccine given each day}

\item{tt_vaccine_doses}{Times at which vaccination rate changes}

\item{duration_of_immunity}{Duration of protection (in days) from vaccines and natural immunity}

\item{duration_of_maternal_immunity}{Duration of maternal immunity}

\item{S_0}{Initial susceptible population}

\item{R_0}{Initial recovered population}

\item{M_0}{Initial maternal immunity population}

\item{additional_parameters}{List of additional parameters}
}
\value{
output array of results
}
\description{
Run model
}
