% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{project_point_estimate}
\alias{project_point_estimate}
\title{Function to project immunity using point estimates of the parameters}
\usage{
project_point_estimate(
  type,
  t_projection_starts,
  t_projection_ends,
  age_group_sizes,
  death_rates,
  tt_death_rates = NULL,
  birth_rates,
  tt_birth_rates = NULL,
  duration_of_maternal_immunity,
  additional_parameters = NULL,
  vaccine_names,
  force_of_infection,
  tt_force_of_infection = NULL,
  vaccine_efficacy,
  vaccine_efficacy_disease,
  vaccinations,
  tt_vaccinations = NULL,
  duration_of_immunity,
  S_0,
  R_0,
  M_0
)
}
\arguments{
\item{type}{Type of model to run}

\item{t_projection_starts}{Time to start projections}

\item{t_projection_ends}{Time to end projections}

\item{age_group_sizes}{Vector of age group sizes}

\item{death_rates}{Vector of death rates}

\item{tt_death_rates}{Vector of time varying death rates}

\item{birth_rates}{Vector of birth rates}

\item{tt_birth_rates}{Vector of time varying birth rates}

\item{duration_of_maternal_immunity}{Duration of maternal immunity}

\item{additional_parameters}{List of additional parameters,default NULL}

\item{vaccine_names}{Vector of vaccine names}

\item{force_of_infection}{list of named vectors of force of infection}

\item{tt_force_of_infection}{list of named vectors of time varying force of infection}

\item{vaccine_efficacy}{list of named vectors of vaccine efficacy}

\item{vaccine_efficacy_disease}{list of named vectors of against disease, should be the total efficacy not adjusted for protection against infection}

\item{vaccinations}{list of named vectors of vaccine dose/rate parameters}

\item{tt_vaccinations}{list of named vectors of time varying vaccine doses}

\item{duration_of_immunity}{list of named vectors of duration of immunity}

\item{S_0}{list of named vectors of initial susceptibles}

\item{R_0}{list of named vectors of initial recovered}

\item{M_0}{list of named vectors of initial maternal immunity}
}
\description{
Runs multiple models for each vaccine type (determined by the first dimension of relevant parameters)
}
